requirements.txt
decord==0.6.0
diffusers==0.20.0
einops==0.7.0
imageio==2.27.0
ipdb==0.13.13
librosa==0.9.2
moviepy==1.0.3
numpy==1.23.5
omegaconf==2.3.0
opencv_python==4.8.0.76
Pillow==10.2.0
pydub==0.25.1
safetensors==0.3.3
scipy==1.12.0
soundfile==0.12.1
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tqdm==4.65.0
transformers==4.32.1
xformers==0.0.23.post1
